move calculate reclamm init amounts to util
