+++ src/utils/index.ts
@@ -11,3 +11,4 @@ export * from './types';
 export * from './balancerV2Contracts';
 export * from './balancerV3Contracts';
 export * from './numbers';
+export * from './calculateReClammInitAmounts';
