+++ test/v3/createPool/reClamm/reClamm.integration.test.ts
@@ -8,7 +8,6 @@ import {
     zeroAddress,
     parseUnits,
     TestActions,
-    parseAbi,
 } from 'viem';
 import {
     CHAINS,
@@ -23,10 +22,7 @@ import {
     vaultExtensionAbi_V3,
     PublicWalletClient,
     InitPoolDataProvider,
-    type InputAmount,
-    type PoolState,
-    fpMulDown,
-    fpDivDown,
+    calculateReClammInitAmounts,
 } from 'src';
 import { ANVIL_NETWORKS, startFork } from '../../../anvil/anvil-global-setup';
 import {
@@ -140,68 +136,21 @@ describe('ReClamm - create & init', () => {
             protocolVersion,
         );
 
-        ///// start of special reclamm init logic /////
-
+        ///// START RECLAMM INIT LOGIC /////
         // User chooses an amount for one of the tokens
         const givenAmountIn = {
-            address: BAL.address,
+            address: BAL.address, // TODO: Also test DAI as given token (can't init same pool twice)
             rawAmount: parseUnits('100', BAL.decimals),
             decimals: BAL.decimals,
         };
 
         // SDK calculates the amount for the other token
-        const amountsIn = await calculateInitReClammAmountsIn(
+        const amountsIn = await calculateReClammInitAmounts({
+            client,
             poolState,
             givenAmountIn,
-        );
-
-        async function calculateInitReClammAmountsIn(
-            poolState: PoolState,
-            givenAmountIn: InputAmount,
-        ): Promise<InputAmount[]> {
-            // fetch proportion value from pool contract
-            const proportion = await client.readContract({
-                address: poolState.address,
-                abi: parseAbi([
-                    'function computeInitialBalanceRatio() external view returns (uint256 balanceRatio)',
-                ]),
-                functionName: 'computeInitialBalanceRatio',
-                args: [],
-            });
-
-            // poolState reads on chain so always has tokens in vault sorted order right?
-            const { tokens } = poolState;
-            const givenTokenIndex = tokens.findIndex(
-                (t) =>
-                    t.address.toLowerCase() ===
-                    givenAmountIn.address.toLowerCase(),
-            );
-
-            let calculatedAmountIn: InputAmount;
-
-            // https://github.com/balancer/reclamm/blob/8207b33c1ab76de3c42b015bab5210a8436376de/test/reClammPool.test.ts#L120-L128
-            if (givenTokenIndex === 0) {
-                // if chosen token is first in sort order, we multiply
-                calculatedAmountIn = {
-                    address: tokens[1].address,
-                    rawAmount: fpMulDown(givenAmountIn.rawAmount, proportion),
-                    decimals: tokens[1].decimals,
-                };
-            } else {
-                // if chosen token is second in sort order, we divide
-                calculatedAmountIn = {
-                    address: tokens[0].address,
-                    rawAmount: fpDivDown(givenAmountIn.rawAmount, proportion),
-                    decimals: tokens[0].decimals,
-                };
-            }
-
-            // Return amounts in consistent order based on token addresses
-            return [givenAmountIn, calculatedAmountIn].sort((a, b) =>
-                a.address.localeCompare(b.address),
-            );
-        }
-        ///// end of special reclamm init logic /////
+        });
+        ///// END RECLAMM INIT LOGIC /////
 
         const initPoolInput = {
             amountsIn,
